Renaming: 'restrictions' -> 'common'
