+++ alcor/services/common.py
