+++ tests/strategies/processing.py
@@ -1,6 +1,6 @@
 from hypothesis import strategies
 
-from alcor.services.restrictions import FILTRATION_METHODS
+from alcor.services.common import FILTRATION_METHODS
 
 filtration_methods = strategies.one_of(map(strategies.just,
                                            FILTRATION_METHODS))
